fix(discord): break channel info cache cycle
